refactor: extract sets list and exercise options and use them in exercise card
